backend/requirement.txt
uvicorn
python-dotenv
fastapi
bs4
requests
numpy
nltk